fix: svg 경로수정
