+++ src/entities/card/config/bank.module.css
