+++ src/entities/card/model/bank.ts
